requirements.txt
pytest==7.1.2; python_version >= "3.7"
pyhamcrest==2.0.3; python_version >= "3.5"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
dataclasses-json===0.5.7
Faker===13.15.1